--- lading_blackholes/Cargo.toml
@@ -1,69 +1,0 @@
-[package]
-name = "lading_blackholes"
-version = "0.6.0"
-authors = ["Brian L. Troutwine <brian@troutwine.us>"]
-edition = "2018"
-license = "MIT"
-repository = "https://github.com/blt/lading/"
-keywords = ["udp", "http", "blackhole"]
-categories = ["command-line-utilities", "development-tools::profiling"]
-description = "Blackhole programs with some instrumentation"
-
-[dependencies]
-tower = { version = "0.4", default-features = false, features = ["timeout", "limit", "load-shed"] }
-metrics = { version = "0.17", default-features = false, features = ["std"] }
-metrics-exporter-prometheus = { version = "0.6", default-features = false, features = ["tokio-exporter"] }
-serde_yaml = "0.8"
-serde_qs = "0.8.5"
-rand = "0.8.4"
-
-[dependencies.argh]
-version = "0.1"
-default-features = false
-features = []
-
-[dependencies.futures]
-version = "0.3.19"
-default-features = false
-features = []
-
-[dependencies.hyper]
-version = "0.14"
-default-features = false
-features = ["http2", "tcp"]
-
-[dependencies.serde]
-version = "1.0"
-default-features = false
-features = ["derive", "std"]
-
-[dependencies.serde_json]
-version = "1.0"
-default-features = false
-features = ["std"]
-
-[dependencies.tokio]
-version = "1.14"
-default-features = false
-features = ["rt", "rt-multi-thread", "macros", "fs", "io-util"]
-
-[dependencies.tokio-util]
-version = "0.6.9"
-default-features = false
-features = ["io"]
-
-[dependencies.once_cell]
-version = "1.8"
-default-features = false
-
-[[bin]]
-name = "tcp_blackhole"
-
-[[bin]]
-name = "udp_blackhole"
-
-[[bin]]
-name = "http_blackhole"
-
-[[bin]]
-name = "sqs_blackhole"
