--- lading_blackholes/README.md
@@ -1,5 +1,0 @@
-# `blackhole` - a collection of blackhole programs
-
-The programs in this sub-project are blackholes. For instance, `udp_blackhole`
-listens on a certain local UDP port, receives packets, makes a note about that
-receipt and then drops everything.
