--- lading_generators/Cargo.toml
@@ -1,90 +1,0 @@
-[package]
-name = "lading_generators"
-readme = "README.md"
-version = "0.7.0"
-authors = ["Brian L. Troutwine <brian@troutwine.us>"]
-edition = "2018"
-license = "MIT"
-repository = "https://github.com/blt/lading/"
-keywords = ["random_test", "generator"]
-categories = ["command-line-utilities", "development-tools::profiling", "filesystem"]
-description = "Generating programs with stable throughput."
-
-[dependencies]
-lading_common = { path = "../lading_common" }
-serde_yaml = "0.8"
-
-[dependencies.argh]
-version = "0.1"
-default-features = false
-features = []
-
-[dependencies.futures]
-version = "0.3"
-default-features = false
-features = ["std"]
-
-[dependencies.byte-unit]
-version = "4.0"
-default-features = false
-features = ["serde"]
-
-[dependencies.http]
-version = "0.2"
-default-features = false
-features = []
-
-[dependencies.http-serde]
-version = "1.0"
-default-features = false
-features = []
-
-[dependencies.hyper]
-version = "0.14"
-default-features = false
-features = ["client", "http2"]
-
-[dependencies.governor]
-version = "0.3"
-default-features = false
-features = ["std", "jitter", "quanta"]
-
-[dependencies.tokio]
-version = "1.14"
-default-features = false
-features = ["rt", "rt-multi-thread", "macros", "fs", "io-util"]
-
-[dependencies.serde]
-version = "1.0"
-default-features = false
-features = ["derive", "std"]
-
-[dependencies.serde_json]
-version = "1.0"
-default-features = false
-features = ["std"]
-
-[dependencies.metrics]
-version = "0.17"
-default-features = false
-features = ["std"]
-
-[dependencies.metrics-exporter-prometheus]
-version = "0.6"
-default-features = false
-features = ["tokio-exporter"]
-
-[dependencies.pretty_env_logger]
-version = "0.4"
-default-features = false
-features = []
-
-[dependencies.rand]
-version = "0.8"
-default-features = false
-features = []
-
-[dependencies.rdkafka]
-version = "0.27.0"
-default-features = false
-features = []
