--- lading_generators/example.yaml
@@ -1,28 +1,0 @@
-# Set the number of background worker threads that this run will use.
-worker_threads: 3
-
-targets:
-  # Create a target 'foo'. This name will be used in telemetry and must be
-  # unique.
-  foo:
-    # Set the path template for this target. %NNN% will be replaced with an
-    # index up to `duplicates - 1`.
-    path_template: "/tmp/playground/%NNN%-foo.json"
-    # Set the total duplicates of this target. This run will make
-    # /tmp/playground/1-foo.json, 2-foo.json and 3-foo.json.
-    duplicates: 3
-    # Set the variant of the lines produced. Available options are 'Json', 'Ascii'
-    # and 'Constant'.
-    variant: "Json"
-    # Set the maximum size of this target's files. When this **soft** limit is
-    # breached the file handler is closed and a new file in the **same**
-    # location is opened. The original file is deleted, simulating a 'rotation'
-    # without increasing the total number of files on disk.
-    maximum_bytes_per_file: "1Gb"
-    # Set the bytes available per second to be written. The target will not be
-    # able to produce bytes any faster than this, though it may chose to produce
-    # bytes less quickly.
-    bytes_per_second: "10Mb"
-    # Set the maximum size of lines that this target will produce. The target
-    # may produce lines of smaller size.
-    maximum_prebuild_cache_size_bytes: "3 Gb"
