--- lading_generators/src/kafka_gen.rs
@@ -1,8 +1,0 @@
-//! The `kafka_gen` library
-//!
-//! This crate is intended to back the `kafka_gen` executable and is
-//! not considered useful otherwise.
-
-pub use worker::Worker;
-pub mod config;
-mod worker;
