--- lading_generators/src/splunk_hec_gen/config.rs
@@ -1,53 +1,0 @@
-//! This module controls configuration parsing from the end user, providing a
-//! convenience mechanism for the rest of the program. Crashes are most likely
-//! to originate from this code, intentionally.
-use std::{collections::HashMap, net::SocketAddr};
-
-use http::Uri;
-use lading_common::payload::SplunkHecEncoding;
-use serde::Deserialize;
-
-/// Main configuration struct for this program
-#[derive(Deserialize, Debug)]
-pub struct Config {
-    /// Total number of worker threads to use in this program
-    pub worker_threads: u16,
-    /// Address and port for prometheus exporter
-    pub prometheus_addr: SocketAddr,
-    /// The [`Target`] instances and their base name
-    pub targets: HashMap<String, Target>,
-}
-
-/// Optional Splunk HEC indexer acknowledgements configuration
-#[derive(Deserialize, Debug)]
-pub struct AckSettings {
-    /// The time in seconds between queries to /services/collector/ack
-    pub ack_query_interval_seconds: u64,
-    /// The time in seconds an ackId can remain pending before assuming data was
-    /// dropped
-    pub ack_timeout_seconds: u64,
-}
-
-/// The [`Target`] instance from which to derive workers
-#[derive(Deserialize, Debug)]
-pub struct Target {
-    /// The seed for random operations against this target
-    pub seed: [u8; 32],
-    /// The URI for the target, must be a valid URI
-    #[serde(with = "http_serde::uri")]
-    pub target_uri: Uri,
-    /// Format used when submitting event data to Splunk HEC
-    pub format: SplunkHecEncoding,
-    /// Splunk HEC authentication token
-    pub token: String,
-    /// Splunk HEC indexer acknowledgements behavior options
-    pub acknowledgements: Option<AckSettings>,
-    /// The maximum size in bytes of the cache of prebuilt messages
-    pub maximum_prebuild_cache_size_bytes: byte_unit::Byte,
-    /// The bytes per second to send or receive from the target
-    pub bytes_per_second: byte_unit::Byte,
-    /// The block sizes for messages to this target
-    pub block_sizes: Option<Vec<byte_unit::Byte>>,
-    /// The total number of parallel connections to maintain
-    pub parallel_connections: u16,
-}
