--- lading_generators/src/tcp_gen.rs
@@ -1,8 +1,0 @@
-//! The `tcp_gen` library
-//!
-//! This crate is intended to back the `tcp_gen` executable and is
-//! not considered useful otherwise.
-
-pub use worker::Worker;
-pub mod config;
-mod worker;
