--- lading_generators/src/tcp_gen/config.rs
@@ -1,43 +1,0 @@
-//! This module controls configuration parsing from the end user, providing a
-//! convenience mechanism for the rest of the program. Crashes are most likely
-//! to originate from this code, intentionally.
-use serde::Deserialize;
-use std::{collections::HashMap, net::SocketAddr};
-
-/// Main configuration struct for this program
-#[derive(Debug, Deserialize)]
-pub struct Config {
-    /// Total number of worker threads to use in this program
-    pub worker_threads: u16,
-    /// Address and port for prometheus exporter
-    pub prometheus_addr: SocketAddr,
-    /// The [`Target`] instances and their base name
-    pub targets: HashMap<String, Target>,
-}
-
-/// Variant of the [`Target`]
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum Variant {
-    /// Generates Fluent messages
-    Fluent,
-    /// Generates syslog5424 messages
-    Syslog5424,
-}
-
-/// The [`Target`] instance from which to derive workers
-#[derive(Debug, Deserialize)]
-pub struct Target {
-    /// The seed for random operations against this target
-    pub seed: [u8; 32],
-    /// The address for the target, must be a valid SocketAddr
-    pub addr: String,
-    /// The payload variant
-    pub variant: Variant,
-    /// The bytes per second to send or receive from the target
-    pub bytes_per_second: byte_unit::Byte,
-    /// The block sizes for messages to this target
-    pub block_sizes: Option<Vec<byte_unit::Byte>>,
-    /// The maximum size in bytes of the cache of prebuilt messages
-    pub maximum_prebuild_cache_size_bytes: byte_unit::Byte,
-}
