+++ lading_rig/src/generator/kafka.rs
@@ -1,5 +1,3 @@
-use crate::kafka_gen::config::Throughput;
-use crate::kafka_gen::config::{Target, Variant};
 use byte_unit::{Byte, ByteUnit};
 use futures::stream::FuturesUnordered;
 use futures::{FutureExt, StreamExt};
@@ -15,8 +13,66 @@ use rdkafka::error::KafkaError;
 use rdkafka::producer::{FutureProducer, FutureRecord};
 use rdkafka::types::RDKafkaErrorCode;
 use rdkafka::ClientConfig;
+use serde::Deserialize;
+use std::collections::HashMap;
 use std::convert::TryInto;
 use std::num::NonZeroU32;
+use tracing::info;
+
+use crate::signals::Shutdown;
+
+/// The throughput configuration
+#[derive(Clone, Copy, Debug, Deserialize)]
+#[serde(rename_all = "snake_case")]
+pub enum Throughput {
+    /// The producer should run as fast as possible.
+    Unlimited,
+    /// The producer is limited to sending a certain number of bytes every second.
+    BytesPerSecond {
+        /// Number of bytes.
+        amount: byte_unit::Byte,
+    },
+    /// The producer is limited to sending a certain number of messages every second.
+    MessagesPerSecond {
+        /// Number of messages.
+        amount: u32,
+    },
+}
+
+/// Variant of the [`TargetTemplate`]
+#[derive(Clone, Copy, Debug, Deserialize)]
+#[serde(rename_all = "snake_case")]
+pub enum Variant {
+    /// Generates Datadog Logs JSON messages
+    DatadogLog,
+    /// Generates a limited subset of FoundationDB logs
+    FoundationDb,
+    /// Generates a line of printable ascii characters
+    Ascii,
+    /// Generates a json encoded line
+    Json,
+}
+
+#[derive(Clone, Debug, Deserialize)]
+pub struct Config {
+    /// The seed for random operations against this target
+    pub seed: [u8; 32],
+    /// Bootstrap server for Kafka.  Used identically like the flag of the same
+    /// name present on Kafka CLI tooling.
+    pub bootstrap_server: String,
+    /// Topic to produce to.
+    pub topic: String,
+    /// The payload generator to use for this target
+    pub variant: Variant,
+    /// The throughput configuration
+    pub throughput: Throughput,
+    /// The maximum size in bytes of the cache of prebuilt messages
+    pub maximum_prebuild_cache_size_bytes: byte_unit::Byte,
+    /// The block sizes for messages to this target
+    pub block_sizes: Option<Vec<byte_unit::Byte>>,
+    /// Map of rdkafka=-specific overrides to apply to the producer
+    pub producer_config: Option<HashMap<String, String>>,
+}
 
 #[derive(Debug)]
 pub enum Error {
@@ -54,14 +110,17 @@ impl From<::std::io::Error> for Error {
 /// The [`Worker`] defines a task that emits variant lines to an HTTP server
 /// controlling throughput.
 #[derive(Debug)]
-pub struct Worker {
-    name: String,
-    target: Target,
+pub struct Kafka {
     block_cache: Vec<Block>,
     labels: Vec<(String, String)>,
+    bootstrap_server: String,
+    topic: String,
+    producer_config: Option<HashMap<String, String>>,
+    throughput: Throughput,
+    shutdown: Shutdown,
 }
 
-impl Worker {
+impl Kafka {
     /// Create a new [`Worker`] instance
     ///
     /// # Errors
@@ -73,13 +132,10 @@ impl Worker {
     /// Function will panic if user has passed non-zero values for any byte
     /// values. Sharp corners.
     #[allow(clippy::cast_possible_truncation)]
-    pub fn new(name: String, target: Target) -> Result<Self, Error> {
-        let labels = vec![
-            ("name".to_string(), name.clone()),
-            ("server".to_string(), target.bootstrap_server.to_string()),
-        ];
+    pub fn new(config: Config, shutdown: Shutdown) -> Result<Self, Error> {
+        let labels = vec![];
 
-        let block_sizes: Vec<usize> = target
+        let block_sizes: Vec<usize> = config
             .block_sizes
             .clone()
             .unwrap_or_else(|| {
@@ -98,18 +154,21 @@ impl Worker {
             .map(|sz| sz.get_bytes() as usize)
             .collect();
         let block_cache = generate_block_cache(
-            target.maximum_prebuild_cache_size_bytes,
-            target.variant,
-            target.seed,
+            config.maximum_prebuild_cache_size_bytes,
+            config.variant,
+            config.seed,
             &block_sizes,
             &labels,
         );
 
         Ok(Self {
-            name,
-            target,
             block_cache,
             labels,
+            bootstrap_server: config.bootstrap_server,
+            producer_config: config.producer_config,
+            throughput: config.throughput,
+            topic: config.topic,
+            shutdown,
         })
     }
 
@@ -122,14 +181,14 @@ impl Worker {
     /// # Panics
     ///
     /// Function will panic if it is unable to produce messages to the target Kafka cluster.
-    pub async fn spin(self) -> Result<(), Error> {
+    pub async fn spin(mut self) -> Result<(), Error> {
         // Configure our Kafka producer.
-        let bootstrap_server = self.target.bootstrap_server;
-        let topic = self.target.topic;
+        let bootstrap_server = self.bootstrap_server;
+        let topic = self.topic;
         let labels = self.labels;
 
         let mut client_config = ClientConfig::new();
-        let mut config_values = self.target.producer_config.unwrap_or_default();
+        let mut config_values = self.producer_config.unwrap_or_default();
         config_values.insert("bootstrap.servers".to_string(), bootstrap_server);
         for (k, v) in config_values.drain() {
             client_config.set(k, v);
@@ -138,8 +197,8 @@ impl Worker {
         let producer = FutureProducer::from_config(&client_config)?;
 
         // Configure our rate limiter.
-        let limit_by_bytes = matches!(self.target.throughput, Throughput::BytesPerSecond { .. });
-        let rate_limiter = get_rate_limiter(self.target.throughput);
+        let limit_by_bytes = matches!(self.throughput, Throughput::BytesPerSecond { .. });
+        let rate_limiter = get_rate_limiter(self.throughput);
 
         let mut in_flight = FuturesUnordered::new();
 
@@ -163,28 +222,35 @@ impl Worker {
             let limiter_n = if limit_by_bytes { block_size.get() } else { 1 };
             let limiter_n = NonZeroU32::new(limiter_n).expect("should never be zero");
 
-            let _ = rate_limiter.until_n_ready(limiter_n).await;
-            let mut record = Some(
-                FutureRecord::to(topic.as_ref())
-                    .payload(&block.bytes)
-                    .key(&()),
-            );
+            tokio::select! {
+                _ = rate_limiter.until_n_ready(limiter_n) => {
+                    let mut record = Some(
+                        FutureRecord::to(topic.as_ref())
+                            .payload(&block.bytes)
+                            .key(&()),
+                    );
 
-            loop {
-                match producer.send_result(record.take().unwrap()) {
-                    Ok(fut) => {
-                        counter!("requests_sent", 1, &labels);
-                        in_flight
-                            .push(async move { fut.await.map(|_| u64::from(block_size.get())) });
-                        break;
-                    }
-                    Err((e, old_record)) => {
-                        if e == KafkaError::MessageProduction(RDKafkaErrorCode::QueueFull) {
-                            record = Some(old_record);
-                            tokio::task::yield_now().await;
+                    loop {
+                        match producer.send_result(record.take().unwrap()) {
+                            Ok(fut) => {
+                                counter!("requests_sent", 1, &labels);
+                                in_flight
+                                    .push(async move { fut.await.map(|_| u64::from(block_size.get())) });
+                                break;
+                            }
+                            Err((e, old_record)) => {
+                                if e == KafkaError::MessageProduction(RDKafkaErrorCode::QueueFull) {
+                                    record = Some(old_record);
+                                    tokio::task::yield_now().await;
+                                }
+                            }
                         }
                     }
                 }
+                _ = self.shutdown.recv() => {
+                    info!("shutdown signal received");
+                    return Ok(());
+                },
             }
         }
     }
