+++ lading_rig/src/lib.rs
@@ -1,5 +1,11 @@
 //! The `rig` library
 
+//#![allow(clippy::multiple_crate_versions)]
+#![deny(clippy::all)]
+#![deny(clippy::cargo)]
+#![deny(clippy::pedantic)]
+//#![deny(missing_docs)]
+
 pub mod blackhole;
 pub mod captures;
 pub mod config;
